+++ src/attachments.rs
@@ -1,12 +1,28 @@
+use crate::{
+    checksum::sha256_digest,
+    console::{print_error, Status},
+    error::Result,
+    responses::Attachment,
+    sync_operation::SyncOperation,
+};
 use std::{
-    io::{self, Write},
+    collections::HashMap,
+    fs::File,
+    io::{self, BufReader, Write},
     path::{Path, PathBuf},
 };
 
+use anyhow::Context;
 use comrak::nodes::NodeLink;
 use regex::Regex;
+use reqwest::blocking::multipart::Part;
 
-use crate::confluence_storage_renderer::{escape_href, WriteWithLast};
+use crate::{
+    confluence_client::ConfluenceClient,
+    confluence_storage_renderer::{escape_href, WriteWithLast},
+    link_generator::LinkGenerator,
+    responses::MultiEntityResult,
+};
 
 #[derive(Debug, PartialEq)]
 pub struct ImageAttachment {
@@ -55,6 +71,97 @@ pub fn render_link_enter(nl: &NodeLink, output: &mut WriteWithLast) -> io::Resul
 
 pub fn render_link_leave(_nl: &NodeLink, output: &mut WriteWithLast) -> io::Result<()> {
     output.write_all(b"</ac:image>")?;
+    Ok(())
+}
+
+pub fn sync_page_attachments(
+    confluence_client: &ConfluenceClient,
+    page_id: &str,
+    page_source: &str,
+    attachments: &[ImageAttachment],
+    link_generator: &mut LinkGenerator,
+) -> Result<()> {
+    let existing_attachments: MultiEntityResult<Attachment> = confluence_client
+        .get_attachments(page_id)?
+        .error_for_status()?
+        .json()?;
+
+    let mut hashes = HashMap::<String, String>::new();
+    let mut remove_titles_to_id = HashMap::<String, String>::new();
+    let mut title_to_fileid = HashMap::<String, String>::new();
+    for existing_attachment in existing_attachments.results.iter() {
+        if existing_attachment.comment.starts_with("hash:") {
+            hashes.insert(
+                existing_attachment.title.clone(),
+                existing_attachment
+                    .comment
+                    .strip_prefix("hash:")
+                    .unwrap()
+                    .into(),
+            );
+        }
+        remove_titles_to_id.insert(
+            existing_attachment.title.clone(),
+            existing_attachment.id.clone(),
+        );
+        title_to_fileid.insert(
+            existing_attachment.title.clone(),
+            existing_attachment.file_id.clone(),
+        );
+    }
+
+    for attachment in attachments.iter() {
+        let attachment_name = attachment.name.clone();
+
+        let id = title_to_fileid[&attachment_name].clone();
+        link_generator.register_attachment_id(page_source, &attachment.url, &id);
+
+        remove_titles_to_id.remove(&attachment_name);
+
+        let op = SyncOperation::start(format!("[{}] attachment", attachment.path.display()), true);
+        let input = File::open(&attachment.path)
+            .with_context(|| format!("Opening attachment for {}", attachment_name))?;
+        let reader = BufReader::new(input);
+        let hashstring = sha256_digest(reader)?;
+        if hashes.contains_key(&attachment_name)
+            && hashstring == *hashes.get(&attachment_name).unwrap()
+        {
+            op.end(Status::Skipped);
+            return Ok(());
+        }
+
+        let file_part = Part::file(&attachment.path)?.file_name(attachment.name.clone());
+
+        let response =
+            confluence_client.create_or_update_attachment(page_id, file_part, &hashstring)?;
+
+        if !response.status().is_success() {
+            // Handle non-2xx responses (e.g., 400 Bad Request)
+            let status = response.status();
+            let error_body = response.text()?;
+            print_error(&format!(
+                "error updating attachment: Status: {}, Body: {}",
+                status, error_body
+            ));
+        }
+
+        op.end(Status::Updated);
+    }
+
+    let _remove_results: Vec<crate::confluence_client::Result> = remove_titles_to_id
+        .iter()
+        .map(|(title, id)| {
+            let op = SyncOperation::start(format!("[{}] attachment", title), false);
+            let result = confluence_client.remove_attachment(id);
+            if result.is_ok() {
+                op.end(Status::Deleted);
+            } else {
+                op.end(Status::Error);
+            }
+            result
+        })
+        .collect();
+
     Ok(())
 }
